requirements.txt
Django==3.1.2
django-allauth==0.43.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework_simplejwt==4.4.0
oauthlib==3.1.0
pathspec==0.5.9
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
django_extensions==3.1.3